refactor: extract @keebmeet/shared workspace for API contract types
Move shared response DTOs, zod validation schemas, and interface types into a
new @keebmeet/shared npm workspace that both backend and frontend depend on,
replacing the frontend's cross-package imports into backend/src. Convert the
repo to npm workspaces and update the Dockerfiles and CI workflows for
workspace-aware builds.
