+++ .github/workflows/test.yml
@@ -30,6 +30,12 @@ jobs:
 
       - name: Install dependencies
         run: npm install
+        working-directory: .
+
+      # Backend tests import @keebmeet/shared, so build it first.
+      - name: Build shared
+        run: npm run build:shared
+        working-directory: .
 
       - name: Run tests
         run: npm run test:ci
