+++ backend/package.json
@@ -31,6 +31,7 @@
     "typescript": "^6.0.0"
   },
   "dependencies": {
+    "@keebmeet/shared": "*",
     "axios": "^1.6.2",
     "bcrypt": "^6.0.0",
     "dayjs": "^1.11.10",
