+++ frontend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "keebmeet",
+  "name": "@keebmeet/frontend",
   "private": true,
   "version": "0.0.0",
   "type": "module",
@@ -9,6 +9,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@keebmeet/shared": "*",
     "@reduxjs/toolkit": "^2.12.0",
     "@tailwindcss/vite": "^4.3.1",
     "axios": "^1.6.1",
