+++ frontend/src/pages/ManageMeetupAttendeesPage.tsx
@@ -21,7 +21,7 @@ import { useMemo, useState, type ReactNode } from 'react';
 import { FiEdit2 } from 'react-icons/fi';
 import { useParams } from 'react-router-dom';
 import { toast } from 'sonner';
-import { type TicketInfo } from '../../../backend/src/controllers/meetups';
+import { type TicketInfo } from '@keebmeet/shared';
 import {
   useEditAttendeeMutation,
   useGetMeetupAttendeesQuery,
