+++ shared/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "@keebmeet/shared",
+  "version": "0.0.0",
+  "private": true,
+  "description": "Shared API contract types and validation schemas for KeebMeet.",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "scripts": {
+    "build": "tsc"
+  },
+  "dependencies": {
+    "zod": "^4.4.3"
+  },
+  "devDependencies": {
+    "typescript": "^6.0.0"
+  }
+}
