+++ shared/src/eventbriteInterfaces.ts
