+++ shared/src/index.ts
@@ -1,0 +1,8 @@
+// Public surface of @keebmeet/shared — the API contract shared by the backend
+// (which produces it) and the frontend (which consumes it).
+export * from './dto';
+export * from './validation';
+export * from './eventbriteInterfaces';
+export * from './meetupInterfaces';
+export * from './rafflesInterfaces';
+export * from './userInterfaces';
