+++ shared/src/meetupInterfaces.ts
