+++ shared/src/rafflesInterfaces.ts
