+++ shared/src/userInterfaces.ts
