+++ shared/src/validation.ts
